requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9          # file uploads

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0                  # async PostgreSQL driver
alembic==1.13.1                  # migrations
psycopg2-binary==2.9.9           # sync driver (for Alembic CLI)

# ── Auth & Security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0 # JWT
passlib[bcrypt]==1.7.4           # password hashing
bcrypt==4.1.3

# ── Validation ────────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ── HTTP / Utilities ──────────────────────────────────────────────────────────
httpx==0.27.0
python-dotenv==1.0.1
aiofiles==23.2.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
pytest-cov==5.0.0
